pyproject.toml
[project]
name = "tp2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
]
